+++ src/ConsoleClient/DB/Data/57.sql
@@ -1,0 +1,5 @@
+START TRANSACTION;
+
+update settings SET value = '57' where name = 'db_version';
+
+COMMIT;