+++ src/ConsoleClient/DB/Structure/57.sql
@@ -1,0 +1,5 @@
+START TRANSACTION;
+
+ALTER TABLE `risks` DROP CONSTRAINT  IF EXISTS `FK_risks_IncidentResponsePlans_IncidentResponsePlanId`;
+
+COMMIT;